+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Date format: DD/MM/YYYY
 
-## [1.0.0] - 31/03/2021
+## [1.0.0] - Initial - 01/04/2021
 
-- Initial Release
+- Get system accent color
+- Check if system is in dark mode