+++ system_theme_web/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.2
+
+- Support for getting accent color
+
 ## 0.0.1
 
-* TODO: Describe initial release.
+- Initial release